# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Quanyi AI Task OS API",
    "version": "0.1.0"
  },
  "paths": {
    "/api/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_api_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Health Api Health Get"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "summary": "Login",
        "operationId": "login_api_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "summary": "Me",
        "operationId": "me_api_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRead"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/projects": {
      "get": {
        "summary": "List Projects",
        "operationId": "list_projects_api_projects_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProjectRead"
                  },
                  "type": "array",
                  "title": "Response List Projects Api Projects Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/projects/{project_id}": {
      "get": {
        "summary": "Get Project",
        "operationId": "get_project_api_projects__project_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks": {
      "get": {
        "summary": "List Tasks",
        "operationId": "list_tasks_api_tasks_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskRead"
                  },
                  "title": "Response List Tasks Api Tasks Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/{task_id}/actions/{action}": {
      "post": {
        "summary": "Task Action",
        "operationId": "task_action_api_tasks__task_id__actions__action__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          },
          {
            "name": "action",
            "in": "path",
            "required": true,
            "schema": {
              "enum": [
                "ACCEPT",
                "START",
                "SUBMIT",
                "APPROVE",
                "RETURN",
                "CANCEL"
              ],
              "type": "string",
              "title": "Action"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskActionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskActionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/{task_id}/submissions": {
      "get": {
        "summary": "Task Submissions",
        "operationId": "task_submissions_api_tasks__task_id__submissions_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SubmissionRead"
                  },
                  "title": "Response Task Submissions Api Tasks  Task Id  Submissions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/{task_id}/history": {
      "get": {
        "summary": "Task History",
        "operationId": "task_history_api_tasks__task_id__history_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StatusHistoryRead"
                  },
                  "title": "Response Task History Api Tasks  Task Id  History Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ExecutionMode": {
        "type": "string",
        "enum": [
          "HUMAN",
          "AI",
          "HYBRID"
        ],
        "title": "ExecutionMode"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest"
      },
      "ProjectRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "client": {
            "type": "string",
            "title": "Client"
          },
          "objective": {
            "type": "string",
            "title": "Objective"
          },
          "owner_id": {
            "type": "string",
            "title": "Owner Id"
          },
          "owner_name": {
            "type": "string",
            "title": "Owner Name"
          },
          "next_milestone": {
            "type": "string",
            "title": "Next Milestone"
          },
          "due_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due At"
          },
          "progress": {
            "type": "integer",
            "title": "Progress"
          },
          "health": {
            "type": "string",
            "enum": [
              "正常",
              "有风险",
              "需关注"
            ],
            "title": "Health"
          },
          "current_stage": {
            "type": "string",
            "title": "Current Stage"
          },
          "stages": {
            "items": {
              "$ref": "#/components/schemas/StageRead"
            },
            "type": "array",
            "title": "Stages"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "client",
          "objective",
          "owner_id",
          "owner_name",
          "next_milestone",
          "due_at",
          "progress",
          "health",
          "current_stage",
          "stages"
        ],
        "title": "ProjectRead"
      },
      "StageRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "position": {
            "type": "integer",
            "title": "Position"
          },
          "status": {
            "$ref": "#/components/schemas/StageStatus"
          },
          "progress": {
            "type": "integer",
            "title": "Progress"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "position",
          "status",
          "progress"
        ],
        "title": "StageRead"
      },
      "StageStatus": {
        "type": "string",
        "enum": [
          "PLANNED",
          "ACTIVE",
          "WAITING_REVIEW",
          "DONE"
        ],
        "title": "StageStatus"
      },
      "StatusHistoryRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "from_status": {
            "type": "string",
            "title": "From Status"
          },
          "to_status": {
            "type": "string",
            "title": "To Status"
          },
          "actor_id": {
            "type": "string",
            "title": "Actor Id"
          },
          "actor_name": {
            "type": "string",
            "title": "Actor Name"
          },
          "action": {
            "type": "string",
            "title": "Action"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "task_id",
          "from_status",
          "to_status",
          "actor_id",
          "actor_name",
          "action",
          "reason",
          "created_at"
        ],
        "title": "StatusHistoryRead"
      },
      "SubmissionRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "submitted_by": {
            "type": "string",
            "title": "Submitted By"
          },
          "submitter_name": {
            "type": "string",
            "title": "Submitter Name"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "external_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Url"
          },
          "asset_reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Asset Reference"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "task_id",
          "version",
          "submitted_by",
          "submitter_name",
          "summary",
          "external_url",
          "asset_reference",
          "created_at"
        ],
        "title": "SubmissionRead"
      },
      "TaskActionRequest": {
        "properties": {
          "expected_version": {
            "type": "integer",
            "title": "Expected Version"
          },
          "summary": {
            "type": "string",
            "title": "Summary",
            "default": ""
          },
          "external_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Url"
          },
          "asset_reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Asset Reference"
          },
          "reason": {
            "type": "string",
            "title": "Reason",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "expected_version"
        ],
        "title": "TaskActionRequest"
      },
      "TaskActionResponse": {
        "properties": {
          "task": {
            "$ref": "#/components/schemas/TaskRead"
          },
          "idempotent_replay": {
            "type": "boolean",
            "title": "Idempotent Replay",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "task"
        ],
        "title": "TaskActionResponse"
      },
      "TaskRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "project_name": {
            "type": "string",
            "title": "Project Name"
          },
          "stage_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "deliverable": {
            "type": "string",
            "title": "Deliverable"
          },
          "acceptance": {
            "type": "string",
            "title": "Acceptance"
          },
          "owner_id": {
            "type": "string",
            "title": "Owner Id"
          },
          "owner_name": {
            "type": "string",
            "title": "Owner Name"
          },
          "reviewer_id": {
            "type": "string",
            "title": "Reviewer Id"
          },
          "reviewer_name": {
            "type": "string",
            "title": "Reviewer Name"
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus"
          },
          "execution_mode": {
            "$ref": "#/components/schemas/ExecutionMode"
          },
          "priority": {
            "type": "string",
            "title": "Priority"
          },
          "progress": {
            "type": "integer",
            "title": "Progress"
          },
          "due_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due At"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "project_name",
          "stage_id",
          "title",
          "description",
          "deliverable",
          "acceptance",
          "owner_id",
          "owner_name",
          "reviewer_id",
          "reviewer_name",
          "status",
          "execution_mode",
          "priority",
          "progress",
          "due_at",
          "source",
          "version"
        ],
        "title": "TaskRead"
      },
      "TaskStatus": {
        "type": "string",
        "enum": [
          "PENDING_OWNER_CONFIRMATION",
          "TODO",
          "IN_PROGRESS",
          "WAITING_EXTERNAL",
          "BLOCKED",
          "WAITING_HUMAN_CONFIRMATION",
          "WAITING_REVIEW",
          "DONE",
          "CANCELED"
        ],
        "title": "TaskStatus"
      },
      "TeamRole": {
        "type": "string",
        "enum": [
          "CEO",
          "MEMBER"
        ],
        "title": "TeamRole"
      },
      "TokenResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In"
          },
          "user": {
            "$ref": "#/components/schemas/UserRead"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "expires_in",
          "user"
        ],
        "title": "TokenResponse"
      },
      "UserRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "role": {
            "$ref": "#/components/schemas/TeamRole"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "name",
          "role"
        ],
        "title": "UserRead"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}